[calamares] Reduce warnings now that getInteger is qint64

+++ src/calamares/CalamaresWindow.cpp
@@ -48,11 +48,11 @@ windowDimensionToPixels( const Calamares::Branding::WindowDimension& u )
     }
     if ( u.unit() == Calamares::Branding::WindowDimensionUnit::Pixies )
     {
-        return u.value();
+        return static_cast< int >( u.value() );
     }
     if ( u.unit() == Calamares::Branding::WindowDimensionUnit::Fonties )
     {
-        return u.value() * CalamaresUtils::defaultFontHeight();
+        return static_cast< int >( u.value() * CalamaresUtils::defaultFontHeight() );
     }
     return 0;
 }
