+++ src/modules/shellprocess/ShellProcessJob.cpp
@@ -69,7 +69,7 @@ void
 ShellProcessJob::setConfigurationMap( const QVariantMap& configurationMap )
 {
     bool dontChroot = CalamaresUtils::getBool( configurationMap, "dontChroot", false );
-    int timeout = CalamaresUtils::getInteger( configurationMap, "timeout", 10 );
+    qint64 timeout = CalamaresUtils::getInteger( configurationMap, "timeout", 10 );
     if ( timeout < 1 )
     {
         timeout = 10;
